chore: move rpcerr pkg into the client pkg
