+++ rpc/call.go
@@ -4,7 +4,7 @@ import (
 	"context"
 
 	"github.com/NethermindEth/juno/core/felt"
-	"github.com/NethermindEth/starknet.go/rpcerr"
+	"github.com/NethermindEth/starknet.go/client/rpcerr"
 )
 
 // Call calls the Starknet Provider's function with the given (Starknet) request and block ID.
