+++ rpc/trace.go
@@ -5,7 +5,7 @@ import (
 	"encoding/json"
 
 	"github.com/NethermindEth/juno/core/felt"
-	"github.com/NethermindEth/starknet.go/rpcerr"
+	"github.com/NethermindEth/starknet.go/client/rpcerr"
 )
 
 // TraceTransaction returns the transaction trace for the given transaction hash.
